prefect/requirements.txt
prefect==2.6.5
requests
pydantic
nats-py